backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose[cryptography]
bcrypt
python-multipart
pydantic
anyio
email-validator
httpx
